size in kb to bytes

+++ src/Application/Interfaces/IStorageManagementService.cs
@@ -2,8 +2,8 @@ namespace GreenRoom.Application.Interfaces;
 
 public interface IStorageManagementService
 {
-    string GenerateUrlToUpload(string fileName, string contentType, int expiryInSeconds);
-    string GenerateUrlToDownload(string fileName, int expiryInSeconds);
+    string GenerateUrlToUpload(string fileName, string contentType, long expiryInSeconds);
+    string GenerateUrlToDownload(string fileName, long expiryInSeconds);
     Task CreateBucketAsync(string bucketName);
     Task CreateFolderAsync(string folderName);
     Task<long> GetBucketSizeAsync();
